Use int instead of long in LERC
git-svn-id: https://svn.osgeo.org/gdal/trunk@33700 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/mrf/libLERC/TImage.hpp
@@ -1,17 +1,23 @@
 /*
 Copyright 2015 Esri
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
+
 http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
+
 A local copy of the license and additional notices are located with the
 source distribution at:
+
 http://github.com/Esri/lerc/
+
 Contributors:  Thomas Maurer
 */
 
@@ -26,9 +32,6 @@ Contributors:  Thomas Maurer
 #include "Image.h"
 
 NAMESPACE_LERC_START
-// -------------------------------------------------------------------------- ;
-// -------------------------------------------------------------------------- ;
-// ---- related classes ----------------------------------------------------- ;
 
 class CntZ
 {
@@ -38,12 +41,6 @@ public:
   bool operator != (const CntZ& cz) const    { return cnt != cz.cnt || z != cz.z; }
   void operator += (const CntZ& cz)          { cnt += cz.cnt;  z += cz.z; }
 };
-
-// -------------------------------------------------------------------------- ;
-
-/**	Template Image
- *	
- */
 
 template< class Element >
 class TImage : public Image
@@ -189,7 +186,7 @@ bool TImage< Element >::operator == (const Image& img) const
 
   const Element* ptr0 = getData();
   const Element* ptr1 = ((const TImage&)img).getData();
-  long cnt = getSize();
+  int cnt = getSize();
   while (cnt--)
     if (*ptr0++ != *ptr1++)
       return false;
